refactor(alert): move ALERT-level track finding from rec.ahdc to rec.alert
The TrackFinding strategy, the Track/TrackCandidate value types, and the AI/* models live in rec.ahdc today, but they are ALERT-level orchestration
across AHDC and ATOF (GNN literally builds a joint AHDC+ATOF graph). They
belong in rec.alert.* alongside the existing rec.alert.TrackMatchingAI,
rec.alert.AIPID, rec.alert.banks, rec.alert.projections.
