+++ reconstruction/alert/src/main/java/org/jlab/rec/ahdc/HoughTransform/HoughTransform.java
@@ -2,7 +2,7 @@ package org.jlab.rec.ahdc.HoughTransform;
 
 import org.jlab.rec.ahdc.AHDCCluster.AHDCCluster;
 import Jama.Matrix;
-import org.jlab.rec.ahdc.Track.TrackCandidate;
+import org.jlab.rec.alert.Track.TrackCandidate;
 
 import java.util.ArrayList;
 import java.util.List;
