+++ reconstruction/alert/src/main/java/org/jlab/rec/ahdc/KalmanFilter/KalmanFilter.java
@@ -14,7 +14,7 @@ import org.jlab.clas.tracking.kalmanfilter.Material;
 import org.jlab.io.base.DataBank;
 import org.jlab.io.base.DataEvent;
 import org.jlab.rec.ahdc.Hit.Hit;
-import org.jlab.rec.ahdc.Track.Track;
+import org.jlab.rec.alert.Track.Track;
 
 //import org.apache.commons.math3.linear.RealMatrixFormat;
