+++ reconstruction/alert/src/main/java/org/jlab/rec/alert/AI/AIPrediction.java
@@ -1,4 +1,4 @@
-package org.jlab.rec.ahdc.AI;
+package org.jlab.rec.alert.AI;
 
 import java.util.ArrayList;
