+++ reconstruction/alert/src/main/java/org/jlab/rec/alert/AI/ModelTrackFinding.java
@@ -1,4 +1,4 @@
-package org.jlab.rec.ahdc.AI;
+package org.jlab.rec.alert.AI;
 
 import ai.djl.MalformedModelException;
 import ai.djl.ndarray.NDArray;
