+++ reconstruction/alert/src/main/java/org/jlab/rec/alert/AI/ModelTrackFindingGNN.java
@@ -1,4 +1,4 @@
-package org.jlab.rec.ahdc.AI;
+package org.jlab.rec.alert.AI;
 
 import java.io.IOException;
 import java.nio.file.Paths;
