+++ reconstruction/alert/src/main/java/org/jlab/rec/alert/AI/TrackPrediction.java
@@ -1,4 +1,4 @@
-package org.jlab.rec.ahdc.AI;
+package org.jlab.rec.alert.AI;
 
 import org.jlab.rec.ahdc.AHDCCluster.AHDCCluster;
 import org.jlab.rec.ahdc.PreCluster.PreCluster;
