+++ reconstruction/alert/src/main/java/org/jlab/rec/alert/Track/Track.java
@@ -1,6 +1,6 @@
-package org.jlab.rec.ahdc.Track;
+package org.jlab.rec.alert.Track;
 
-import org.jlab.rec.ahdc.AI.InterCluster;
+import org.jlab.rec.alert.AI.InterCluster;
 import org.jlab.rec.ahdc.AHDCCluster.AHDCCluster;
 import org.jlab.rec.ahdc.HelixFit.HelixFitObject;
 import org.jlab.rec.ahdc.Hit.Hit;
