+++ reconstruction/alert/src/main/java/org/jlab/rec/alert/TrackFinding/DistanceTrackFinder.java
@@ -1,18 +1,18 @@
-package org.jlab.rec.ahdc.TrackFinding;
+package org.jlab.rec.alert.TrackFinding;
 
 import org.jlab.rec.ahdc.AHDCCluster.AHDCCluster;
 import org.jlab.rec.ahdc.AHDCCluster.AHDCClusterFinder;
+import org.jlab.rec.ahdc.Distance.Distance;
 import org.jlab.rec.ahdc.Hit.Hit;
-import org.jlab.rec.ahdc.HoughTransform.HoughTransform;
 import org.jlab.rec.ahdc.PreCluster.PreCluster;
 import org.jlab.rec.ahdc.PreCluster.PreClusterFinder;
-import org.jlab.rec.ahdc.Track.AtofHitStub;
-import org.jlab.rec.ahdc.Track.TrackCandidate;
+import org.jlab.rec.alert.Track.AtofHitStub;
+import org.jlab.rec.alert.Track.TrackCandidate;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class HoughTrackFinder implements TrackFinder {
+public class DistanceTrackFinder implements TrackFinder {
 
     @Override
     public TrackFinderResult findTracks(ArrayList<Hit> hits, List<AtofHitStub> atofHits) {
@@ -24,9 +24,9 @@ public class HoughTrackFinder implements TrackFinder {
         cf.findCluster(preclusters);
         ArrayList<AHDCCluster> clusters = cf.get_AHDCClusters();
 
-        HoughTransform hough = new HoughTransform();
-        hough.find_tracks(clusters);
-        ArrayList<TrackCandidate> tracks = hough.get_AHDCTrackCandidates();
+        Distance distance = new Distance();
+        distance.find_track(clusters);
+        ArrayList<TrackCandidate> tracks = distance.get_AHDCTrackCandidates();
 
         return TrackFinderResult.ok(tracks);
     }
