+++ reconstruction/alert/src/main/java/org/jlab/service/alert/ALERTEngine.java
@@ -27,7 +27,7 @@ import org.jlab.rec.alert.AIPID.ModelPrePID;
 import org.jlab.rec.alert.banks.RecoBankWriter;
 import org.jlab.rec.alert.projections.TrackProjector;
 import org.jlab.rec.atof.hit.ATOFHit;
-import org.jlab.rec.ahdc.AI.InterCluster;
+import org.jlab.rec.alert.AI.InterCluster;
 import org.jlab.rec.ahdc.AHDCCluster.AHDCCluster;
 import org.jlab.rec.ahdc.DocaCluster.DocaCluster;
 import org.jlab.rec.ahdc.DocaCluster.DocaClusterRefiner;
@@ -37,17 +37,17 @@ import org.jlab.rec.ahdc.TrackFindingMode;
 import org.jlab.rec.ahdc.PreCluster.PreCluster;
 import org.jlab.rec.ahdc.PreCluster.PreClusterFinder;
 import org.jlab.rec.ahdc.Hit.Hit;
-import org.jlab.rec.ahdc.TrackFinding.AITrackFinder;
-import org.jlab.rec.ahdc.TrackFinding.DistanceTrackFinder;
-import org.jlab.rec.ahdc.TrackFinding.GNNTrackFinder;
-import org.jlab.rec.ahdc.TrackFinding.HoughTrackFinder;
-import org.jlab.rec.ahdc.TrackFinding.TrackFinder;
-import org.jlab.rec.ahdc.TrackFinding.TrackFinderResult;
+import org.jlab.rec.alert.TrackFinding.AITrackFinder;
+import org.jlab.rec.alert.TrackFinding.DistanceTrackFinder;
+import org.jlab.rec.alert.TrackFinding.GNNTrackFinder;
+import org.jlab.rec.alert.TrackFinding.HoughTrackFinder;
+import org.jlab.rec.alert.TrackFinding.TrackFinder;
+import org.jlab.rec.alert.TrackFinding.TrackFinderResult;
 import org.jlab.geom.detector.alert.AHDC.AlertDCDetector;
 import org.jlab.geom.detector.alert.AHDC.AlertDCFactory;
-import org.jlab.rec.ahdc.Track.AtofHitStub;
-import org.jlab.rec.ahdc.Track.Track;
-import org.jlab.rec.ahdc.Track.TrackCandidate;
+import org.jlab.rec.alert.Track.AtofHitStub;
+import org.jlab.rec.alert.Track.Track;
+import org.jlab.rec.alert.Track.TrackCandidate;
 import org.jlab.clas.pdg.PDGDatabase;
 import org.jlab.clas.pdg.PDGParticle;
 import java.util.List;
